Remove dead code and consolidate tmux packages
- Remove internal/tmux wrapper package, migrate all imports to pkg/tmux
- Remove unused functions from internal/bugreport/collector.go:
  GetRedactor(), RedactPath(), RedactedLogPath()
- Remove unused RepoNotFound() from internal/errors/errors.go
- Remove unused functions from internal/format/format.go:
  Success(), Warning(), Error(), Info(), Table.Print(), BranchName()
- Fix deprecated filepath.HasPrefix usage in pkg/config/config_test.go

Total: -662 lines of dead code removed.
